Removing Dependency on Gulp and Added Notes
Created a new server.js file that contains a function for starting the server. The gulpfile imports that function to start the server, so the original gulp task stays intact. The build folder is being added to the project so that the project can be run without gulp. The server can now be started with 'gulp' or 'npm start'. Closes #1.

Some additional scripts and notes were also added, since having people copy and paste them from other sites just adds overhead to this walkthrough.